requirements.txt
biopython>=1.68
numpy>=1.11.2
pandas>=0.19.1
#pkg-resources==0.0.0
python-dateutil==2.6.0
pytz==2016.10
requests==2.12.4
six==1.10.0
tqdm==4.10.0
